# api-gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_handler.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: study.khoipn.id.vn
info:
  contact:
    email: support@study.khoipn.id.vn
    name: Study Platform Support
    url: https://study.khoipn.id.vn/support
  description: API Gateway for Study Platform - A comprehensive learning management
    system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://study.khoipn.id.vn/terms
  title: Study Platform API
  version: "1.0"
paths:
...
  /payments/purchase/course/{course_id}:
    post:
      consumes:
      - application/json
      description: Purchase a course with payment method
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Purchase successful
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Purchase course
      tags:
      - Payments
...
  /payments/stripe/payment-intents:
    post:
      consumes:
      - application/json
      description: Create Stripe payment intent for course purchase
      produces:
      - application/json
      responses:
        "201":
          description: Payment intent created
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Create payment intent
      tags:
      - Payments
  /payments/stripe/payment-intents/{payment_intent_id}:
    get:
      description: Get payment intent details by ID
      parameters:
      - description: Payment Intent ID
        in: path
        name: payment_intent_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Payment intent details
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Get payment intent
      tags:
      - Payments
  /payments/stripe/transactions:
    get:
      description: List user's Stripe payment transactions
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of transactions
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: List Stripe transactions
      tags:
      - Payments
...
  /payments/transactions:
    get:
      description: List user's payment transactions
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of transactions
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: List transactions
      tags:
      - Payments
...
produces:
- application/json
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...